requirements.txt
livekit
livekit-agents
livekit-plugins-google
livekit-plugins-turn-detector
python-dotenv
google-auth
google-api-python-client
flask
